Moved searchbar section to own component, makes home a server component again
